New extension test for confiming extension activation and command registration.
